+++ src/adaptive_interpolator.cpp
@@ -39,7 +39,7 @@ tree2::interpolator::Interpolator
 test_adaptive_interpolator(Rcpp::Function f, double a, double b) {
   tree2::util::RFunctionWrapper fw(f);
   const double atol = 1e-6, rtol = 1e-6;
-  const int nbase = 17, max_depth = 16;
+  const size_t nbase = 17, max_depth = 16;
   tree2::interpolator::AdaptiveInterpolator
     generator(atol, rtol, nbase, max_depth);
   return generator.construct(fw, a, b);
